+++ src/opts/images.rs
@@ -32,6 +32,12 @@ pub enum NetworkMode {
     Custom(String),
 }
 
+impl Default for NetworkMode {
+    fn default() -> Self {
+        NetworkMode::Bridge
+    }
+}
+
 impl AsRef<str> for NetworkMode {
     fn as_ref(&self) -> &str {
         match self {
@@ -478,6 +484,12 @@ impl AsRef<str> for PullPolicy {
             PullPolicy::Newer => "newer",
             PullPolicy::Never => "never",
         }
+    }
+}
+
+impl Default for PullPolicy {
+    fn default() -> Self {
+        PullPolicy::Always
     }
 }