+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * Add `SystemdUnitsOptsBuilder::requires`
 * Add `SystemdUnitsOptsBuilder::wants`
 * Rename `SocketNotifyMode::ConmonOnly` -> `SocketNotifyMode::Conmon` and fix it's key when serialized
+* Add default implementations for `DiffType`, `RestartPolicy`, `ImageVolumeMode`, `SeccompPolicy`, `SystemdEnabled`, `NetworkMode`, `PullPolicy`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync